Complete first part of homework
