+++ oop/src/main/java/com/dudka/courses/shapes/Shape.java
@@ -1,4 +1,4 @@
-package com.dudka.courses;
+package com.dudka.courses.shapes;
 
 public interface Shape {
      double volume();
